Finally!!!!

+++ User/Drive/inc/USART.h
@@ -15,8 +15,8 @@ typedef enum {
 } UARTS;
 
 typedef struct{
-    volatile uint8_t Start_Counter;
-    volatile uint8_t End_Counter;
+    volatile uint16_t Start_Counter;
+    volatile uint16_t End_Counter;
     uint8_t RX_Buffer[RX_BUFFER_LEN];
 } UART_Buffer;
 
@@ -31,7 +31,7 @@ void UART_Init(UARTS UART_Select,u32 baudrate);
 void UART_Send_Data(UARTS UART_Select,u8 Data);
 void UART_Clear_Buffer(UARTS UART_Select);
 u8 UART_Get_Data(UARTS UART_Select);
-u8 UART_Get_Length(UARTS UART_Select);
+u16 UART_Get_Length(UARTS UART_Select);
 void UART_Send_Array(UARTS UART_Select,u8 *Array,int Length);
 void UART_Set_baudrate(UARTS UART_Select,u32 baudrate);
 u8 UART_Get_Data_With_Position(UARTS UART_Select, int position);
